:recycle:: Server: Refactor server_t removing client structure

+++ server/src/map_init.c
@@ -30,7 +30,7 @@ static map_t *create_map(server_t *server)
 static int nb_total(server_t *server)
 {
     int mapValue = server->params->x * server->params->y;
-    int density[7] = {0.5, 0.3, 0.15, 0.1, 0.1, 0.08, 0.05};
+    float density[7] = {0.5, 0.3, 0.15, 0.1, 0.1, 0.08, 0.05};
     int total = ((mapValue * density[0]) + (mapValue * density[1]) + (mapValue
         * density[2]) + (mapValue * density[3]) + (mapValue * density[4]) +
         (mapValue * density[5]) + (mapValue * density[6]));
@@ -68,7 +68,7 @@ static void malloc_ressource(server_t *server, int j, int i, tiles_t *tiles)
 static void init_tiles(server_t *server)
 {
     int mapV = server->params->x * server->params->y;
-    int den[7] = {0.5, 0.3, 0.15, 0.1, 0.1, 0.08, 0.05};
+    float den[7] = {0.5, 0.3, 0.15, 0.1, 0.1, 0.08, 0.05};
     int nb[7] = {(mapV * den[0]), (mapV * den[1]), (mapV * den[2]), (mapV *
         den[3]), (mapV * den[4]), (mapV * den[5]), (mapV * den[6])};
     tiles_t *tiles = shuffle_fisher(server->params->x, server->params->y);
